+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/event/UserDatasetUninstallEvent.java
@@ -5,7 +5,7 @@ import java.util.Set;
 import org.gusdb.wdk.model.user.dataset.UserDatasetType;
 
 public class UserDatasetUninstallEvent extends UserDatasetEvent {
-  UserDatasetUninstallEvent(Integer eventId, Set<String> projectsFilter, Integer userDatasetId, UserDatasetType userDatasetType) {
+  UserDatasetUninstallEvent(Long eventId, Set<String> projectsFilter, Integer userDatasetId, UserDatasetType userDatasetType) {
     super(eventId, projectsFilter, userDatasetId, userDatasetType);
   }
 }
